log ardupilot: Use ringbuffer to hold data in memory - flush as needed
Avoid using `lseek` to place message on correct position: now, a
ringbuffer is used. When buffer is full, it writes *half* of data to
file - this allows to get some messages out of order.
Naturally, when stop logging, all data on ringbuffer is flushed to file.
This ringbuffer uses naive calculation - it could be improved by fixing
buffer size as a power of two and using masking on operations to avoid
some branching.